Consolidate TimesketchExporter and TimesketchExporterThreaded
