+++ data/recipes/aws_turbinia_ts.json
@@ -72,7 +72,7 @@
       }
     }, {
       "wants": ["TurbiniaGCPProcessorThreaded"],
-      "name": "TimesketchExporterThreaded",
+      "name": "TimesketchExporter",
       "args": {
         "incident_id": "@incident_id",
         "token_password": "@token_password",
