+++ data/recipes/gcp_turbinia_ts_threaded.json
@@ -22,7 +22,7 @@
         }
     }, {
         "wants": ["TurbiniaGCPProcessorThreaded"],
-        "name": "TimesketchExporterThreaded",
+        "name": "TimesketchExporter",
         "args": {
             "incident_id": "@incident_id",
             "token_password": "@token_password",
