+++ dftimewolf/lib/exporters/timesketch.py
@@ -1,22 +1,24 @@
 # -*- coding: utf-8 -*-
-"""Export processing results to Timesketch."""
+"""Export processing results to Timesketch.
+Threaded version of existing Timesketch module."""
 
 import re
 import time
-from typing import Optional, List
+from typing import Optional, List, Type
 
 from timesketch_import_client import importer
 from timesketch_api_client import sketch as ts_sketch
 from timesketch_api_client import client as ts_client  # pylint: disable=unused-import,line-too-long  # used for typing
 
 from dftimewolf.lib import module, timesketch_utils
-from dftimewolf.lib.containers import containers
+from dftimewolf.lib.containers import containers, interface
 from dftimewolf.lib.modules import manager as modules_manager
 from dftimewolf.lib.state import DFTimewolfState
 
 
-class TimesketchExporter(module.BaseModule):
-  """Exports a given set of plaso or CSV files to Timesketch.
+class TimesketchExporter(module.ThreadAwareModule):
+  """Exports a given set of plaso or CSV files to Timesketch. This is a
+  threaded version of an equivalent module.
 
   input: A list of paths to plaso or CSV files.
   output: A URL to the generated timeline.
@@ -42,6 +44,8 @@ class TimesketchExporter(module.BaseModule):
     self.timesketch_api = None  # type: ts_client.TimesketchApi
     self._analyzers = []  # type: List[str]
     self.wait_for_timelines = False
+    self.host_url = None
+    self.sketch = None  #type: ts_sketch
 
   def SetUp(self,  # pylint: disable=arguments-differ
             incident_id: None=None,
@@ -65,7 +69,7 @@ class TimesketchExporter(module.BaseModule):
       wait_for_timelines (bool): Whether to wait until timelines are processed
           in the Timesketch server or not.
     """
-    self.wait_for_timelines = bool(wait_for_timelines)
+    self.wait_for_timelines = wait_for_timelines
 
     self.timesketch_api = timesketch_utils.GetApiClient(
         self.state, token_password=token_password)
@@ -74,8 +78,8 @@ class TimesketchExporter(module.BaseModule):
           'Unable to get a Timesketch API client, try deleting the files '
           '~/.timesketchrc and ~/.timesketch.token', critical=True)
     self.incident_id = incident_id
-    self.sketch_id = int(sketch_id) if sketch_id else 0
-    sketch = None
+    self.sketch_id = sketch_id if sketch_id else 0
+    self.sketch = None
 
     # Check that we have a timesketch session.
     if not (self.timesketch_api or self.timesketch_api.session):
@@ -89,19 +93,19 @@ class TimesketchExporter(module.BaseModule):
 
     # If we have a sketch ID, check that we can write to it and cache it.
     if self.sketch_id:
-      sketch = self.timesketch_api.get_sketch(self.sketch_id)
-      if 'write' not in sketch.my_acl:
+      self.sketch = self.timesketch_api.get_sketch(self.sketch_id)
+      if 'write' not in self.sketch.my_acl:
         self.ModuleError(
             'No write access to sketch ID {0:d}, aborting'.format(
                 self.sketch_id),
             critical=True)
-      self.state.AddToCache('timesketch_sketch', sketch)
-      self.sketch_id = sketch.id
+      self.state.AddToCache('timesketch_sketch', self.sketch)
+      self.sketch_id = self.sketch.id
 
     if analyzers and isinstance(analyzers, (tuple, list)):
       self._analyzers = analyzers
 
-  def _CreateSketch(self, incident_id: None=None) -> ts_sketch.Sketch:
+  def _CreateSketch(self, incident_id: Optional[str]=None) -> ts_sketch.Sketch:
     """Creates a new Timesketch sketch.
 
     Args:
@@ -152,67 +156,49 @@ class TimesketchExporter(module.BaseModule):
           return sketch_id
     return 0
 
-  def Process(self) -> None:
-    """Executes a Timesketch export."""
-    if not self.timesketch_api:
-      message = 'Could not connect to Timesketch server'
-      self.ModuleError(message, critical=True)
+  def Process(self, container: containers.File) -> None:
+    """Executes a Timesketch export.
 
-    sketch = self.state.GetFromCache('timesketch_sketch')
-    if not sketch and self.sketch_id:
-      self.logger.info('Using exiting sketch: {0:d}'.format(self.sketch_id))
-      sketch = self.timesketch_api.get_sketch(self.sketch_id)
-
-    # Create the sketch if no sketch was stored in the cache.
-    if not sketch:
-      sketch = self._CreateSketch(incident_id=self.incident_id)
-      self.sketch_id = sketch.id
-      self.logger.info('New sketch created: {0:d}'.format(self.sketch_id))
+    Args:
+      container (containers.File): A container holding a File to import."""
 
     recipe_name = self.state.recipe.get('name', 'no_recipe')
-    input_names = []
 
-    for file_container in self.state.GetContainers(containers.File):
-      description = file_container.name
-      if not description:
-        continue
+    description = container.name
+    if description:
       name = description.rpartition('.')[0]
+      name = name if name else description
       name = name.replace(' ', '_').replace('-', '_')
-      input_names.append(name)
-
-    if input_names:
       timeline_name = '{0:s}_{1:s}'.format(
-          recipe_name, '_'.join(input_names))
+          recipe_name, name)
     else:
       timeline_name = recipe_name
 
     with importer.ImportStreamer() as streamer:
-      streamer.set_sketch(sketch)
+      streamer.set_sketch(self.sketch)
       streamer.set_timeline_name(timeline_name)
 
-      for file_container in self.state.GetContainers(containers.File):
-        path = file_container.path
-        streamer.add_file(path)
-        if streamer.response and file_container.description:
-          streamer.timeline.description = file_container.description
+      path = container.path
+      streamer.add_file(path)
+      if streamer.response and container.description:
+        streamer.timeline.description = container.description
 
-    api_root = sketch.api.api_root
-    host_url = api_root.partition('api/v1')[0]
-    sketch_url = '{0:s}sketches/{1:d}/'.format(host_url, sketch.id)
+    api_root = self.sketch.api.api_root
+    self.host_url = api_root.partition('api/v1')[0]
+    sketch_url = '{0:s}sketches/{1:d}/'.format(self.host_url, self.sketch.id)
     message = 'Your Timesketch URL is: {0:s}'.format(sketch_url)
-    self.logger.success(message)
-    container = containers.Report(
+    report_container = containers.Report(
         module_name='TimesketchExporter',
         text=message,
         text_format='markdown')
-    self.state.StoreContainer(container)
+    self.state.StoreContainer(report_container)
 
     if self.wait_for_timelines:
       self.logger.info('Waiting for timelines to finish processing...')
       self._WaitForTimelines()
 
     for analyzer in self._analyzers:
-      results = sketch.run_analyzer(
+      results = self.sketch.run_analyzer(
           analyzer_name=analyzer, timeline_name=timeline_name)
       if not results:
         self.logger.info('Analyzer [{0:s}] not able to run on {1:s}'.format(
@@ -227,6 +213,41 @@ class TimesketchExporter(module.BaseModule):
       self.logger.info('Analyzer: {0:s} is running, session ID: {1:d}'.format(
           analyzer, session_id))
       self.logger.info(results.status_string)
+
+  @staticmethod
+  def GetThreadOnContainerType() -> Type[interface.AttributeContainer]:
+    return containers.File
+
+  def GetThreadPoolSize(self) -> int:
+    return 5
+
+  def PreSetUp(self) -> None:
+    pass
+
+  def PostSetUp(self) -> None:
+    pass
+
+  def PreProcess(self) -> None:
+    """Get the sketch, creating it if it doesn't yet exist."""
+    if not self.timesketch_api:
+      message = 'Could not connect to Timesketch server'
+      self.ModuleError(message, critical=True)
+
+    self.sketch = self.state.GetFromCache('timesketch_sketch')
+    if not self.sketch and self.sketch_id:
+      self.logger.info('Using exiting sketch: {0:d}'.format(self.sketch_id))
+      self.sketch = self.timesketch_api.get_sketch(self.sketch_id)
+
+    # Create the sketch if no sketch was stored in the cache.
+    if not self.sketch:
+      self.sketch = self._CreateSketch(incident_id=self.incident_id)
+      self.sketch_id = self.sketch.id
+      self.logger.info('New sketch created: {0:d}'.format(self.sketch_id))
+
+  def PostProcess(self) -> None:
+    sketch_url = '{0:s}sketches/{1:d}/'.format(self.host_url, self.sketch.id)
+    message = 'Your Timesketch URL is: {0:s}'.format(sketch_url)
+    self.logger.success(message)
 
 
 modules_manager.ModulesManager.RegisterModule(TimesketchExporter)
