+++ tests/lib/exporters/timesketch.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-"""Tests the Timesketch exporter."""
+"""Tests the Threaded Timesketch exporter."""
 
 import unittest
 
@@ -9,12 +9,12 @@ import mock
 from dftimewolf import config
 from dftimewolf.lib import state
 from dftimewolf.lib.containers import containers
-from dftimewolf.lib.exporters import timesketch
+from dftimewolf.lib.exporters import timesketch as timesketch
 from dftimewolf.lib import errors
 
 
 class TimesketchExporterTest(unittest.TestCase):
-  """Tests for the Timesketch exporter."""
+  """Tests for the Threaded Timesketch exporter."""
 
   def testInitialization(self):
     """Tests that the processor can be initialized."""
@@ -26,9 +26,7 @@ class TimesketchExporterTest(unittest.TestCase):
   @mock.patch('dftimewolf.lib.timesketch_utils.GetApiClient')
   def testSetupWithSketchId(self, mock_GetApiClient):
     """Tests the SetUp function."""
-    mock_sketch = mock.Mock()
-    mock_sketch.id = 1234
-    mock_sketch.my_acl = ['write']
+    mock_sketch = mock.Mock(id=1234, my_acl=['write'])
     mock_api_client = mock.Mock()
     mock_api_client.get_sketch.return_value = mock_sketch
     mock_GetApiClient.return_value = mock_api_client
@@ -46,9 +44,7 @@ class TimesketchExporterTest(unittest.TestCase):
   @mock.patch('dftimewolf.lib.timesketch_utils.GetApiClient')
   def testSetupWithReadonlySketchId(self, mock_GetApiClient):
     """Tests the SetUp function."""
-    mock_sketch = mock.Mock()
-    mock_sketch.id = 1234
-    mock_sketch.my_acl = ['read']
+    mock_sketch = mock.Mock(id=1234, my_acl=['read'])
     mock_api_client = mock.Mock()
     mock_api_client.get_sketch.return_value = mock_sketch
     mock_GetApiClient.return_value = mock_api_client
@@ -69,9 +65,7 @@ class TimesketchExporterTest(unittest.TestCase):
   @mock.patch('dftimewolf.lib.timesketch_utils.GetApiClient')
   def testNewSketchCreation(self, mock_GetApiClient, _):
     """Tests the SetUp function."""
-    mock_sketch = mock.Mock()
-    mock_sketch.id = 1234
-    mock_sketch.my_acl = ['write']
+    mock_sketch = mock.Mock(id=1234, my_acl=['write'])
     mock_sketch.api.api_root = 'timesketch.com/api/v1'
     mock_api_client = mock.Mock()
     mock_api_client.get_sketch.return_value = None
@@ -85,7 +79,10 @@ class TimesketchExporterTest(unittest.TestCase):
         sketch_id=None,
         analyzers=None
     )
-    timesketch_exporter.Process()
+    test_container = containers.File('file.ext', '/tmp/file.ext')
+    timesketch_exporter.PreProcess()
+    timesketch_exporter.Process(test_container)
+    timesketch_exporter.PostProcess()
     self.assertEqual(timesketch_exporter.sketch_id, 1234)
     mock_api_client.create_sketch.assert_called_with(
       'Untitled sketch', 'Sketch generated by dfTimewolf')
@@ -103,9 +100,7 @@ class TimesketchExporterTest(unittest.TestCase):
   def testWaitForTimeline(
         self, mock_GetApiClient, unused_streamer, unused_sleep):
     """Tests the SetUp function."""
-    mock_sketch = mock.Mock()
-    mock_sketch.id = 1234
-    mock_sketch.my_acl = ['write']
+    mock_sketch = mock.Mock(id=1234, my_acl=['write'])
     mock_sketch.api.api_root = 'timesketch.com/api/v1'
 
     mock_api_client = mock.Mock()
@@ -130,7 +125,10 @@ class TimesketchExporterTest(unittest.TestCase):
         analyzers=None,
         wait_for_timelines=True
     )
-    timesketch_exporter.Process()
+    test_container = containers.File('file.ext', '/tmp/file.ext')
+    timesketch_exporter.PreProcess()
+    timesketch_exporter.Process(test_container)
+    timesketch_exporter.PostProcess()
     mock_sketch.list_timelines.assert_called_once()
